+++ src/JSInterop/Microsoft.JSInterop/src/Infrastructure/DotNetObjectReferenceManager.cs
@@ -5,27 +5,27 @@ using System;
 using System.Collections.Concurrent;
 using System.Threading;
 
-namespace Microsoft.JSInterop
+namespace Microsoft.JSInterop.Infrastructure
 {
     internal class DotNetObjectReferenceManager
     {
         private long _nextId = 0; // 0 signals no object, but we increment prior to assignment. The first tracked object should have id 1
-        private readonly ConcurrentDictionary<long, IDotNetObjectRef> _trackedRefsById = new ConcurrentDictionary<long, IDotNetObjectRef>();
+        private readonly ConcurrentDictionary<long, IDotNetObjectReference> _trackedRefsById = new ConcurrentDictionary<long, IDotNetObjectReference>();
 
         public static DotNetObjectReferenceManager Current
         {
             get
             {
-                if (!(JSRuntime.Current is JSRuntime jsRuntimeBase))
+                if (!(JSRuntime.Current is JSRuntime jsRuntime))
                 {
-                    throw new InvalidOperationException("JSRuntime must be set up correctly and must be an instance of JSRuntimeBase to use DotNetObjectRef.");
+                    throw new InvalidOperationException("JSRuntime must be set up correctly and must be an instance of JSRuntimeBase to use DotNetObjectReference.");
                 }
 
-                return jsRuntimeBase.ObjectRefManager;
+                return jsRuntime.ObjectRefManager;
             }
         }
 
-        public long TrackObject(IDotNetObjectRef dotNetObjectRef)
+        public long TrackObject(IDotNetObjectReference dotNetObjectRef)
         {
             var dotNetObjectId = Interlocked.Increment(ref _nextId);
             _trackedRefsById[dotNetObjectId] = dotNetObjectRef;
@@ -33,7 +33,7 @@ namespace Microsoft.JSInterop
             return dotNetObjectId;
         }
 
-        public IDotNetObjectRef FindDotNetObject(long dotNetObjectId)
+        public IDotNetObjectReference FindDotNetObject(long dotNetObjectId)
         {
             return _trackedRefsById.TryGetValue(dotNetObjectId, out var dotNetObjectRef)
                 ? dotNetObjectRef
